feat: add BankConnectionPolicy to manage user permissions for bank connections
