+++ app/Policies/ConnectedAccountPolicy.php
@@ -31,17 +31,9 @@ final class ConnectedAccountPolicy
     /**
      * Determine whether the user can create models.
      */
-    public function create(User $user): bool
+    public function create(): bool
     {
-        if ($user->is_admin || $user->subscribed('business') || $user->subscribed('enterprise')) {
-            return true;
-        }
-
-        if ($user->subscribed('personal')) {
-            return $user->accounts()->count() < 1;
-        }
-
-        return false;
+        return true;
     }
 
     /**
